+++ frontend/src/components/course/BackendGrading.tsx
@@ -1,0 +1,51 @@
+import { useState } from 'react'
+import { apiFetch } from '../../services/api'
+
+function BackendGrading({ assignments, onSave }: { assignments: any[]; onSave: (sid: string, letter: string) => void }) {
+  const [assignmentId, setAssignmentId] = useState<string>('')
+  const [items, setItems] = useState<any[]>([])
+  const [grades, setGrades] = useState<Record<string, string>>({})
+  const load = async (id: string) => {
+    if (!id) { setItems([]); return }
+    const data = await apiFetch<{ submissions: any[] }>(`/api/assignments/${id}/submissions`)
+    setItems(data.submissions || [])
+  }
+  const save = async (sid: string) => {
+    const letter = grades[sid]
+    if (!letter) return
+    const map: Record<string, number> = { A: 95, B: 85, C: 75, D: 65, E: 55, F: 45 }
+    await apiFetch('/api/submissions/grade', { method: 'POST', body: { submission_id: Number(sid), score: map[letter] ?? 0, feedback: null } })
+    onSave(sid, letter)
+  }
+  return (
+    <div>
+      <div style={{ display: 'flex', gap: 8, alignItems: 'center', marginBottom: 8 }}>
+        <select className="select" value={assignmentId} onChange={(e) => { setAssignmentId(e.target.value); void load(e.target.value) }}>
+          <option value="">Select assignment</option>
+          {assignments.map((a: any) => (<option key={a.id} value={a.id}>{a.title}</option>))}
+        </select>
+      </div>
+      {items.length === 0 ? <p className="muted">No submissions to grade.</p> : (
+        <ul className="list">
+          {items.map((s) => (
+            <li key={s.id}>
+              {s.files?.[0]?.filename || 'file'} — {s.student_name || s.student_email}
+              <select className="select" style={{ marginLeft: 8 }} value={grades[s.id] ?? ''} onChange={(e) => setGrades((m) => ({ ...m, [s.id]: e.target.value }))}>
+                <option value="">Select grade</option>
+                <option value="A">A</option>
+                <option value="B">B</option>
+                <option value="C">C</option>
+                <option value="D">D</option>
+                <option value="E">E</option>
+                <option value="F">F</option>
+              </select>
+              <button className="btn btn-primary" style={{ marginLeft: 8 }} onClick={() => void save(s.id)}>Save</button>
+            </li>
+          ))}
+        </ul>
+      )}
+    </div>
+  )
+}
+
+export default BackendGrading
