+++ frontend/src/components/course/MenuTiny.tsx
@@ -1,0 +1,15 @@
+import { useState } from 'react'
+
+export default function MenuTiny({ onDelete }: { onDelete: () => void }) {
+  const [open, setOpen] = useState(false)
+  return (
+    <div style={{ position: 'relative' }}>
+      <button className="btn" onClick={(e)=>{ e.stopPropagation(); setOpen((v)=>!v) }} aria-label="More">⋮</button>
+      {open && (
+        <div className="card" style={{ position: 'absolute', right: 0, marginTop: 4, zIndex: 10 }}>
+          <button className="btn" onClick={(e) => { e.stopPropagation(); setOpen(false); onDelete() }}>Delete</button>
+        </div>
+      )}
+    </div>
+  )
+}
